+++ test/api/mps.jl
@@ -56,7 +56,7 @@ end
 	# while `DMRG1` requires a scheme carrying the bond dimension `D`
 	alg5 = SVDCompression(truncdim(chi))
 	alg6 = DMRG1(truncdim(chi), initguess=:svd)
-	alg7 = SVDCompression(trunccutoff(ϵ=1.0e-10))
+	alg7 = SVDCompression(truncrelerr(ϵ=1.0e-10))
 	algs = [alg1, alg2, alg3, alg4, alg5, alg6, alg7]
 	tol = 1.0e-7
 	for (name, randmps) in MPSConstructors